requirements.txt
matplotlib~=3.5.0
numpy~=1.21.5
pyserial~=3.5
pyqtgraph~=0.12.3
pyqt5~=5.15.6
pygame~=2.1.2
tendo~=0.2.15
pillow~=8.4.0
pandas~=1.4.1
pyinstaller~=5.0.1